Perform comprehensive code quality audit and professionalize naming.
- Downgrade Java target version to 21 for Maven projects to fix build errors.
- Rename offensive project and package names (AppDelPorcoDio/AuradelPorDios -> AuraApp/AuraProject).
- Refactor repetitive string literals and logic into constants and helper methods.
- Fix empty catch blocks by adding stack trace printing.
- Verify full repository build stability.
- Add CODE_QUALITY_REPORT.md.
