--- Laboratorio/JavaFX/AppDelPorcoDio/src/main/java/aura/appdelporcodio/HelloApplication.java
@@ -1,19 +1,0 @@
-package aura.appdelporcodio;
-
-import javafx.application.Application;
-import javafx.fxml.FXMLLoader;
-import javafx.scene.Scene;
-import javafx.stage.Stage;
-
-import java.io.IOException;
-
-public class HelloApplication extends Application {
-    @Override
-    public void start(Stage stage) throws IOException {
-        FXMLLoader fxmlLoader = new FXMLLoader(HelloApplication.class.getResource("main_view.fxml"));
-        Scene scene = new Scene(fxmlLoader.load(), 600, 400);
-        stage.setTitle("App del porco dio");
-        stage.setScene(scene);
-        stage.show();
-    }
-}
