--- Laboratorio/JavaFX/AppDelPorcoDio/src/main/java/aura/appdelporcodio/Launcher.java
@@ -1,9 +1,0 @@
-package aura.appdelporcodio;
-
-import javafx.application.Application;
-
-public class Launcher {
-    public static void main(String[] args) {
-        Application.launch(HelloApplication.class, args);
-    }
-}
