--- Laboratorio/JavaFX/AppDelPorcoDio/src/main/java/module-info.java
@@ -1,8 +1,0 @@
-module aura.appdelporcodio {
-    requires javafx.controls;
-    requires javafx.fxml;
-
-
-    opens aura.appdelporcodio to javafx.fxml;
-    exports aura.appdelporcodio;
-}
