+++ Laboratorio/JavaFX/AuraApp/src/main/java/aura/auraapp/Launcher.java
@@ -1,0 +1,9 @@
+package aura.auraapp;
+
+import javafx.application.Application;
+
+public class Launcher {
+    public static void main(String[] args) {
+        Application.launch(HelloApplication.class, args);
+    }
+}
