+++ Laboratorio/JavaFX/AuraApp/src/main/java/module-info.java
@@ -1,0 +1,8 @@
+module aura.auraapp {
+    requires javafx.controls;
+    requires javafx.fxml;
+
+
+    opens aura.auraapp to javafx.fxml;
+    exports aura.auraapp;
+}
