+++ Laboratorio/JavaFX/AuraApp/src/main/resources/aura/auraapp/hello-view.fxml
