+++ Laboratorio/JavaFX/AuraApp/src/main/resources/aura/auraapp/main_view.fxml
