+++ Laboratorio/JavaFX/AuraProject/src/main/java/aura/auraproject/HelloApplication.java
@@ -1,0 +1,19 @@
+package aura.auraproject;
+
+import javafx.application.Application;
+import javafx.fxml.FXMLLoader;
+import javafx.scene.Scene;
+import javafx.stage.Stage;
+
+import java.io.IOException;
+
+public class HelloApplication extends Application {
+    @Override
+    public void start(Stage stage) throws IOException {
+        FXMLLoader fxmlLoader = new FXMLLoader(HelloApplication.class.getResource("arena-view.fxml"));
+        Scene scene = new Scene(fxmlLoader.load(), 900, 700);
+        stage.setTitle("Aura Arena dei Maghi");
+        stage.setScene(scene);
+        stage.show();
+    }
+}
