+++ Laboratorio/JavaFX/AuraProject/src/main/java/aura/auraproject/HelloController.java
@@ -1,0 +1,14 @@
+package aura.auraproject;
+
+import javafx.fxml.FXML;
+import javafx.scene.control.Label;
+
+public class HelloController {
+    @FXML
+    private Label welcomeText;
+
+    @FXML
+    protected void onHelloButtonClick() {
+        welcomeText.setText("Welcome to JavaFX Application!");
+    }
+}
