+++ Laboratorio/JavaFX/AuraProject/src/main/java/aura/auraproject/Launcher.java
@@ -1,0 +1,9 @@
+package aura.auraproject;
+
+import javafx.application.Application;
+
+public class Launcher {
+    public static void main(String[] args) {
+        Application.launch(HelloApplication.class, args);
+    }
+}
