+++ Laboratorio/JavaFX/AuraProject/src/main/java/aura/auraproject/arena/SpellType.java
@@ -1,0 +1,7 @@
+package aura.auraproject.arena;
+
+public enum SpellType {
+    ATTACCO,
+    CURA,
+    SHELL_PAYLOAD
+}
