+++ Laboratorio/JavaFX/AuraProject/src/main/java/module-info.java
@@ -1,0 +1,8 @@
+module aura.auraproject {
+    requires javafx.controls;
+    requires javafx.fxml;
+
+
+    opens aura.auraproject to javafx.fxml;
+    exports aura.auraproject;
+}
