--- Laboratorio/JavaFX/AuradelPorDios/src/main/java/aura/auradelpordios/HelloController.java
@@ -1,14 +1,0 @@
-package aura.auradelpordios;
-
-import javafx.fxml.FXML;
-import javafx.scene.control.Label;
-
-public class HelloController {
-    @FXML
-    private Label welcomeText;
-
-    @FXML
-    protected void onHelloButtonClick() {
-        welcomeText.setText("Welcome to JavaFX Application!");
-    }
-}
