--- Laboratorio/JavaFX/AuradelPorDios/src/main/java/aura/auradelpordios/Launcher.java
@@ -1,9 +1,0 @@
-package aura.auradelpordios;
-
-import javafx.application.Application;
-
-public class Launcher {
-    public static void main(String[] args) {
-        Application.launch(HelloApplication.class, args);
-    }
-}
