--- Laboratorio/JavaFX/AuradelPorDios/src/main/java/aura/auradelpordios/arena/SpellType.java
@@ -1,7 +1,0 @@
-package aura.auradelpordios.arena;
-
-public enum SpellType {
-    ATTACCO,
-    CURA,
-    SHELL_PAYLOAD
-}
