--- Laboratorio/JavaFX/AuradelPorDios/src/main/java/module-info.java
@@ -1,8 +1,0 @@
-module aura.auradelpordios {
-    requires javafx.controls;
-    requires javafx.fxml;
-
-
-    opens aura.auradelpordios to javafx.fxml;
-    exports aura.auradelpordios;
-}
