--- Laboratorio/JavaFX/AuradelPorDios/src/main/resources/aura/auradelpordios/arena-view.fxml
@@ -1,61 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.geometry.Insets?>
-<?import javafx.scene.control.Button?>
-<?import javafx.scene.control.Label?>
-<?import javafx.scene.control.TextArea?>
-<?import javafx.scene.image.ImageView?>
-<?import javafx.scene.layout.BorderPane?>
-<?import javafx.scene.layout.HBox?>
-<?import javafx.scene.layout.VBox?>
-<?import javafx.scene.text.Font?>
-
-<BorderPane fx:id="mainPane" prefHeight="750.0" prefWidth="1000.0" style="-fx-background-color: #ff00ff;" xmlns="http://javafx.com/javafx/21" xmlns:fx="http://javafx.com/fxml/1" fx:controller="aura.auradelpordios.ArenaController">
-   <top>
-      <VBox alignment="CENTER" spacing="5.0" BorderPane.alignment="CENTER">
-         <BorderPane.margin>
-            <Insets top="10.0" />
-         </BorderPane.margin>
-         <children>
-            <Label fx:id="titleLabel" text="🔥🔥 AURA MAXIMA 9000 🔥🔥" textFill="#00ff00">
-               <font>
-                  <Font name="Comic Sans MS Bold" size="50.0" />
-               </font>
-            </Label>
-            <Label text="L'ARENA DEI MEGA SMONGOLI" textFill="#ffff00">
-               <font>
-                  <Font name="Comic Sans MS Italic" size="28.0" />
-               </font>
-            </Label>
-            <Button fx:id="auraInfinitaButton" mnemonicParsing="false" onAction="#onAuraInfinita" style="-fx-background-color: gold; -fx-border-color: red; -fx-border-width: 5; -fx-text-fill: red;" text="🚨 ATTIVA AURA INFINITA 🚨">
-               <font>
-                  <Font name="Impact" size="24.0" />
-               </font>
-            </Button>
-         </children>
-      </VBox>
-   </top>
-   <center>
-      <HBox fx:id="wizardsContainer" alignment="CENTER" spacing="15.0" BorderPane.alignment="CENTER">
-         <BorderPane.margin>
-            <Insets bottom="10.0" left="10.0" right="10.0" top="10.0" />
-         </BorderPane.margin>
-      </HBox>
-   </center>
-   <bottom>
-      <VBox alignment="CENTER" spacing="10.0" BorderPane.alignment="CENTER">
-         <BorderPane.margin>
-            <Insets bottom="15.0" left="15.0" right="15.0" />
-         </BorderPane.margin>
-         <children>
-            <TextArea fx:id="battleLog" editable="false" prefHeight="200.0" prefWidth="950.0" style="-fx-control-inner-background: #000000; -fx-text-fill: #00ff00; -fx-font-family: 'Comic Sans MS'; -fx-font-size: 16px; -fx-font-weight: bold; -fx-border-color: #00ffff; -fx-border-width: 4;" wrapText="true" />
-            <HBox alignment="CENTER" spacing="40.0">
-               <children>
-                  <Button fx:id="nextTurnButton" mnemonicParsing="false" onAction="#onNextTurn" style="-fx-background-color: #00ffff; -fx-text-fill: #ff00ff; -fx-font-family: 'Comic Sans MS'; -fx-font-weight: bold; -fx-font-size: 22px; -fx-border-color: black; -fx-border-width: 2;" text="👉 SKIPPA TURNO 👈" />
-                  <Button fx:id="autoBattleButton" mnemonicParsing="false" onAction="#onAutoBattle" style="-fx-background-color: #ff0000; -fx-text-fill: #ffff00; -fx-font-family: 'Comic Sans MS'; -fx-font-weight: bold; -fx-font-size: 22px; -fx-border-color: white; -fx-border-width: 2;" text="🤖 BOTTALA TUTTA 🤖" />
-               </children>
-            </HBox>
-         </children>
-      </VBox>
-   </bottom>
-</BorderPane>
